# contracts/openapi/platform-draft.yaml
openapi: 3.1.0
info:
  title: OpsWeave platform API
  version: 0.1.0-draft
  description: >
    Identity-scoped business endpoints. Tenant and permissions are taken from the
    trusted principal, never from query or body. Explicit OIDC mode uses the authorization-code BFF;
    dev mode accepts only its configured loopback credential. Cookie requests require exact Origin
    and X-CSRF-TOKEN on unsafe methods; GET session metadata supplies the masked token.
    Zabbix Host ingest is available only when an explicit source mode is configured.
    All API responses use Cache-Control no-store and an X-OpsWeave-Request-Id UUID.
    The optional client request ID is correlation metadata only, not identity or idempotency.
paths:
...
  /api/v1/integrations/workflows/log-streams/workflows/{id}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
      - {name: revision, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 10000}}
    get:
      operationId: getWorkflowLogStream
      description: Read bounded log stream task and confirmed batch metadata for one owner-scoped workflow version.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Log stream status
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-log-stream-status.schema.json'}
        default: {description: 400 invalid revision; 401 unauthenticated; 403 denied; 404 hidden workflow; 503 stream unavailable}
  /api/v1/integrations/workflows/log-streams/commands/{id}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getWorkflowLogStreamReceipt
      description: Read the original log stream control receipt; it never repeats a command.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Log stream receipt
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-log-stream-receipt.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 404 hidden receipt; 503 stream unavailable}
...
  /api/v1/integrations/workflows/log-streams/workflows/{id}/verification:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
    post:
      operationId: verifyWorkflowLogStream
      description: Confirm one bounded log batch by id and return the refreshed stream status.
      security: [{bearerAuth: []}, {browserSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '../schemas/v2/workflow-log-stream-verification.schema.json'}
      responses:
        '200':
          description: Updated log stream status
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-log-stream-status.schema.json'}
        default: {description: 400 invalid batch; 401 unauthenticated; 403 denied; 404 hidden batch; 409 already confirmed; 503 stream unavailable}
  /api/v1/integrations/workflows/log-streams/workflows/{id}/batches/{batchId}/records:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
      - {name: batchId, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getWorkflowLogStreamRecords
      description: Read bounded confirmed log records for one metric stream batch. The server checks owner scope and batch state before reading the configured store.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Stored log stream records
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-log-stream-data.schema.json'}
        default: {description: 400 invalid path; 401 unauthenticated; 403 denied; 404 hidden batch; 409 batch not confirmed; 503 store unavailable}
  /api/v1/integrations/workflows/log-streams/workflows/{id}/batches/{batchId}/records/after/{afterIndex}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
      - {name: batchId, in: path, required: true, schema: {type: string, format: uuid}}
      - {name: afterIndex, in: path, required: true, schema: {type: integer, minimum: 0, maximum: 999}}
    get:
      operationId: pageWorkflowLogStreamRecords
      description: Read a bounded continuation page of confirmed log stream records after an explicit nonnegative index.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Stored log stream records page
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-log-stream-data.schema.json'}
        default: {description: 400 invalid index; 401 unauthenticated; 403 denied; 404 hidden batch; 409 batch not confirmed; 503 store unavailable}
  /api/v1/integrations/workflows/metric-outputs:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
    get:
      operationId: getWorkflowMetricOutputCapability
      description: Read the fixed metric sample output capability and server budget. It does not read or write a metric batch.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Metric output capability
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-metric-output-capability.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 503 unavailable}
    post:
      operationId: writeWorkflowMetricOutput
      description: Persist one bounded, owner-authorized metric sample using a pinned preview. Idempotency is based on the request id and no source is contacted.
      security: [{bearerAuth: []}, {browserSession: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '../schemas/v2/workflow-metric-output-request.schema.json'}
      responses:
        '200':
          description: Original metric output receipt
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-metric-output-receipt.schema.json'}
        default: {description: 400 invalid sample; 401 unauthenticated; 403 denied; 409 stale preview or reused request; 503 unavailable}
  /api/v1/integrations/workflows/metric-outputs/commands/{id}:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: getWorkflowMetricOutputReceipt
      description: Read the original metric output receipt after an uncertain response; it never repeats the write.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Metric output receipt
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-metric-output-receipt.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 404 hidden receipt; 503 unavailable}
  /api/v1/integrations/workflows/metric-outputs/workflows/{id}/receipts:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, pattern: '^[a-z][a-z0-9_-]{0,47}$'}}
    get:
      operationId: pageWorkflowMetricOutputReceipts
      description: Read a bounded owner-scoped history of metric output receipts.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Metric output history
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-metric-output-history.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 404 hidden workflow; 503 unavailable}
  /api/v1/integrations/workflows/metric-outputs/commands/{id}/verification:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: verifyWorkflowMetricOutput
      description: Read the server-side verification result for one metric output receipt.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Metric output receipt verification
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-metric-output-receipt.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 404 hidden receipt; 503 unavailable}
  /api/v1/integrations/workflows/metric-outputs/commands/{id}/points:
    parameters:
      - {$ref: '#/components/parameters/clientRequestId'}
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    get:
      operationId: readWorkflowMetricOutputPoints
      description: Read bounded stored metric points after receipt verification. Raw source payloads are not returned.
      security: [{bearerAuth: []}, {browserSession: []}]
      responses:
        '200':
          description: Stored metric points
          content:
            application/json:
              schema: {$ref: '../schemas/v2/workflow-metric-output-points.schema.json'}
        default: {description: 401 unauthenticated; 403 denied; 404 hidden receipt; 409 verification incomplete; 503 unavailable}
...
components:
  parameters:
    clientRequestId:
      in: header
      name: X-OpsWeave-Request-Id
      required: false
      description: Optional single UUID; malformed or duplicate values return 400. Server generates one if absent and echoes a canonical UUID on every API response. Never grants permissions or suppresses duplicate writes.
      schema: {type: string, format: uuid}
      example: 6c8d4385-c206-4b59-8c2e-f9d34f8b3500
...
  securitySchemes:
    historyService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dedicated OAuth client_credentials access token profile described in ../history-service-identity.md. Never accepted as a browser or Runtime credential.
    runtimeKey:
      type: apiKey
      in: header
      name: X-OpsWeave-Runtime-Key
      description: Explicit local process attestation only; does not grant user permissions and is never exposed to Web/model.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Explicit loopback dev credential, or short-lived server-only Runtime delegation on allowlisted endpoints. Raw provider bearer JWTs are not accepted.
    browserSession:
      type: apiKey
      in: cookie
      name: __Host-opsweave
      description: OIDC BFF HttpOnly Secure SameSite=Lax host cookie. Explicit loopback protocol tests use opsweave-test-session. Unsafe methods also require X-CSRF-TOKEN and the configured Origin.